added exiting queue functionality
